+++ Changelog.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 - APAGenerator: Arreglado error con capitulo de libro
+- Authors: Al eliminar un autor también se elimina del storage
 
 ## [1.7.6] - 2024-04-04